+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Changelog
 
-## Next version
+## 0.5.7 - January 13, 2025
 
 ### ✨ Improved
 
 * Return the differential voltage in `read_ion_pumps`.
 * Updated IP addresses for ion pumps.
 * `Retrier` now accepts a `timeout` parameter.
+
 
 ## 0.5.6 - January 11, 2025